Establish MQTT connection to the broker
