+++ android/app/build.gradle
@@ -7,8 +7,8 @@ android {
     buildToolsVersion "30.0.3"
 
     defaultConfig {
-        applicationId "com.example.android"
-        minSdkVersion 16
+        applicationId "com.groupe17.smartdelirover"
+        minSdkVersion 28
         targetSdkVersion 30
         versionCode 1
         versionName "1.0"
