+++ android/app/src/main/java/com/example/android/MainActivity.java
@@ -1,45 +1,102 @@
 package com.example.android;
 
 import androidx.appcompat.app.AppCompatActivity;
-import android.view.View;
-import android.widget.ImageView;
-import android.content.Intent;
 import android.os.Bundle;
-import android.widget.Button;
+import android.util.Log;
+import android.widget.Toast;
 
+import org.eclipse.paho.client.mqttv3.IMqttActionListener;
+import org.eclipse.paho.client.mqttv3.IMqttDeliveryToken;
+import org.eclipse.paho.client.mqttv3.IMqttToken;
+import org.eclipse.paho.client.mqttv3.MqttCallback;
+import org.eclipse.paho.client.mqttv3.MqttMessage;
+//MQTT connection part of the code is built on https://github.com/DIT112-V21/smartcar-mqtt-controller.git
+public class MainActivity extends AppCompatActivity {
+    private static final String TAG = "SmartcarMqttController";
+    private static final String EXTERNAL_MQTT_BROKER = "aerostun.dev";
+    private static final String LOCALHOST = "10.0.2.2";
+    private static final String MQTT_SERVER = "tcp://" + EXTERNAL_MQTT_BROKER + ":1883";
+    private static final String THROTTLE_CONTROL = "/smartcar/control/throttle";
+    private static final String STEERING_CONTROL = "/smartcar/control/steering";
 
-public class MainActivity extends AppCompatActivity implements View.OnClickListener{
-    private ImageView mCameraView;
+    private static final int QOS = 1;
+
+    private MqttClient mMqttClient;
+    private boolean isConnected = false;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-        mCameraView = findViewById(R.id.imageView);
+        mMqttClient = new MqttClient(getApplicationContext(), MQTT_SERVER, TAG);
 
-        Button login_button = findViewById(R.id.login_button);
-        Button signup_button = findViewById(R.id.signup_button);
-
-        login_button.setOnClickListener(this);
-        signup_button.setOnClickListener(this);
+        connectToMqttBroker();
     }
 
     @Override
-    public void onClick(View v) {
-        switch (v.getId()) {
-            case R.id.login_button:
-                openLogin();
-                break;
-            case R.id.signup_button:
-                openSignup();
-                break;
+    protected void onResume() {
+        super.onResume();
+        connectToMqttBroker();
+    }
+
+    @Override
+    protected void onPause() {
+        super.onPause();
+
+        mMqttClient.disconnect(new IMqttActionListener() {
+            @Override
+            public void onSuccess(IMqttToken asyncActionToken) {
+                Log.i(TAG, "Disconnected from broker");
+            }
+
+            @Override
+            public void onFailure(IMqttToken asyncActionToken, Throwable exception) {
+            }
+        });
+    }
+
+    private void connectToMqttBroker() {
+        if (!isConnected) {
+            mMqttClient.connect(TAG, "", new IMqttActionListener() {
+                @Override
+                public void onSuccess(IMqttToken asyncActionToken) {
+                    isConnected = true;
+
+                    final String successfulConnection = "Connected to MQTT broker";
+                    Log.i(TAG, successfulConnection);
+                    Toast.makeText(getApplicationContext(), successfulConnection, Toast.LENGTH_SHORT).show();
+
+                    mMqttClient.subscribe("/smartcar/ultrasound/front", QOS, null);
+
+                }
+
+                @Override
+                public void onFailure(IMqttToken asyncActionToken, Throwable exception) {
+                    final String failedConnection = "Failed to connect to MQTT broker";
+                    Log.e(TAG, failedConnection);
+                    Toast.makeText(getApplicationContext(), failedConnection, Toast.LENGTH_SHORT).show();
+                }
+            }, new MqttCallback() {
+                @Override
+                public void connectionLost(Throwable cause) {
+                    isConnected = false;
+
+                    final String connectionLost = "Connection to MQTT broker lost";
+                    Log.w(TAG, connectionLost);
+                    Toast.makeText(getApplicationContext(), connectionLost, Toast.LENGTH_SHORT).show();
+                }
+
+                @Override
+                public void messageArrived(String topic, MqttMessage message) throws Exception {
+
+                    Log.i(TAG, "[MQTT] Topic: " + topic + " | Message: " + message.toString());
+                }
+
+                @Override
+                public void deliveryComplete(IMqttDeliveryToken token) {
+                    Log.d(TAG, "Message delivered");
+                }
+            });
         }
-    }
-    public void openLogin() {
-        Intent intent = new Intent(this, LoginActivity.class);
-        startActivity(intent);
-    }
-    public void openSignup() {
-        Intent intent = new Intent(this, SignupActivity.class);
-        startActivity(intent);
     }
 }
